#192 Adding more missing error handling

+++ src/main/java/org/swisspush/redisques/action/DeleteAllQueueItemsAction.java
@@ -1,5 +1,6 @@
 package org.swisspush.redisques.action;
 
+import io.vertx.core.AsyncResult;
 import io.vertx.core.Vertx;
 import io.vertx.core.eventbus.Message;
 import io.vertx.core.json.JsonObject;
@@ -32,27 +33,40 @@ public class DeleteAllQueueItemsAction extends AbstractQueueAction {
         JsonObject payload = event.body().getJsonObject(PAYLOAD);
         boolean unlock = payload.getBoolean(UNLOCK, false);
         String queue = payload.getString(QUEUENAME);
-
-        redisProvider.redis().onSuccess(redisAPI ->
-                        redisAPI.del(Collections.singletonList(buildQueueKey(queue)))
-                                .onSuccess(deleteReply -> {
-                                    queueStatisticsCollector.resetQueueFailureStatistics(queue, (Throwable ex, Void v) -> {
-                                        if (ex != null) log.warn("TODO_2958iouhj error handling", ex);
-                                    });
-                                    if (unlock) {
-                                        redisAPI.hdel(Arrays.asList(locksKey, queue)).onSuccess(
-                                                        unlockReply -> handleDeleteQueueReply(event, deleteReply))
-                                                .onFailure(throwable -> handleFail(event, "Failed to unlock queue " + queue, throwable));
-                                    } else {
-                                        handleDeleteQueueReply(event, deleteReply);
-                                    }
-                                })
-                                .onFailure(throwable -> handleFail(event, "Operation DeleteAllQueueItems failed", throwable)))
-                .onFailure(throwable -> handleFail(event, "Operation DeleteAllQueueItems failed", throwable));
-
+        var p = redisProvider.redis();
+        p.onSuccess(redisAPI -> {
+            redisAPI.del(Collections.singletonList(buildQueueKey(queue)), deleteReply -> {
+                if (deleteReply.failed()) {
+                    handleFail(event, "Operation DeleteAllQueueItems failed", deleteReply.cause());
+                    return;
+                }
+                queueStatisticsCollector.resetQueueFailureStatistics(queue, (Throwable ex, Void v) -> {
+                    if (ex != null) log.warn("TODO_2958iouhj error handling", ex);
+                });
+                if (unlock) {
+                    redisAPI.hdel(Arrays.asList(locksKey, queue), unlockReply -> {
+                        if (unlockReply.failed()) {
+                            handleFail(event, "Failed to unlock queue " + queue, unlockReply.cause());
+                        } else {
+                            handleDeleteQueueReply(event, deleteReply);
+                        }
+                    });
+                } else {
+                    handleDeleteQueueReply(event, deleteReply);
+                }
+            });
+        });
+        p.onFailure(ex -> {
+            handleFail(event, "Operation DeleteAllQueueItems failed", ex);
+        });
     }
 
-    private void handleDeleteQueueReply(Message<JsonObject> event, Response reply) {
-        event.reply(createOkReply().put(VALUE, reply.toLong()));
+    private void handleDeleteQueueReply(Message<JsonObject> event, AsyncResult<Response> reply) {
+        if (reply.succeeded()) {
+            event.reply(createOkReply().put(VALUE, reply.result().toLong()));
+        } else {
+            log.error("Failed to replyResultGreaterThanZero", exceptionFactory.newException(reply.cause()));
+            event.reply(createErrorReply());
+        }
     }
 }
